requirements.txt
langchain==1.0.5
langchain-core==1.0.4
langchain-openai==1.0.2
langchain-community==0.4.1

openai>=1.0.0
tiktoken>=0.7.0
pydantic>=2.8.0
python-dotenv>=1.0.1
ddgs>=1.0.0
wikipedia>=1.4.0

